refacto: introduce typescript
