+++ jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  collectCoverageFrom: ["src/*"],
+  preset: "ts-jest",
+  setupFiles: ["./setupJest.js"],
+  testEnvironment: "node",
+  verbose: true
+};
