+++ tsconfig.json
@@ -10,6 +10,7 @@
     "moduleResolution": "node",
     "noUnusedLocals": true,
     "noUnusedParameters": true,
+    "resolveJsonModule": true,
     "sourceMap": true,
     "strict": true,
     "target": "es2018"
